backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
faiss-cpu>=1.7.0
pytest>=7.0.0
